+++ manuscript/code/multi/lib/MyAPI.pm
@@ -1,0 +1,9 @@
+package MyAPI;
+use Dancer2;
+
+get '/hello' => sub {
+    header( 'Content-Type'  => 'application/json' );
+    return encode_json { result => 'Hello API' };
+};
+
+1;